✏️ [fix] 여행 회원 정산 종료 API 요청 경로 변경
trip -> trip-member로 요청 경로 변경
