+++ src/main/java/com/lets/go/right/now/domain/trip/controller/TripController.java
@@ -1,12 +1,7 @@
 package com.lets.go.right.now.domain.trip.controller;
 
 import com.lets.go.right.now.domain.trip.service.TripService;
-import com.lets.go.right.now.global.jwt.dto.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -15,11 +10,4 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class TripController {
     private final TripService tripService;
-
-    @PutMapping("{trip_id}/settlement-finish")
-    public ResponseEntity<?> finishSettlement(
-            @AuthenticationPrincipal CustomUserDetails customUserDetails,
-            @PathVariable("trip_id") Long tripId) {
-        return tripService.finishSettlement(customUserDetails.getEmail(), tripId);
-    }
 }
