--- src/main/java/com/lets/go/right/now/domain/trip/enums/Status.java
@@ -1,5 +1,0 @@
-package com.lets.go.right.now.domain.trip.enums;
-
-public enum Status {
-    PROGRESS, DONE
-}
