+++ src/main/java/com/lets/go/right/now/domain/trip/service/TripService.java
@@ -1,7 +1,4 @@
 package com.lets.go.right.now.domain.trip.service;
 
-import org.springframework.http.ResponseEntity;
-
 public interface TripService {
-    ResponseEntity<?> finishSettlement(String email, Long tripId);
 }
