+++ src/main/java/com/lets/go/right/now/domain/tripMember/enums/Status.java
@@ -1,0 +1,5 @@
+package com.lets.go.right.now.domain.tripMember.enums;
+
+public enum Status {
+    PROGRESS, DONE
+}
