+++ src/main/java/com/lets/go/right/now/domain/tripMember/service/TripMemberService.java
@@ -1,0 +1,7 @@
+package com.lets.go.right.now.domain.tripMember.service;
+
+import org.springframework.http.ResponseEntity;
+
+public interface TripMemberService {
+    ResponseEntity<?> finishSettlement(String email, Long tripId);
+}
